# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CWCM API
  version: 0.1.1-draft
  description: >
    Draft API contract baseline for Corporate Wallpaper Campaign Manager.
    This document reflects the intended backend surface required by the PRD.
    Every bearer-protected operation may return the standard 401 error response;
    the reverse proxy may return a structured 503 API_UNAVAILABLE response when
    the application server cannot be reached.
servers:
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /wallpapers:
    get:
      tags: [Wallpapers]
      summary: List wallpapers
      description: Operator or Administrator only; includes uploader and all campaign references.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallpaper list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Wallpaper"
    post:
      tags: [Wallpapers]
      summary: Upload wallpaper
      description: Operator or Administrator only. Accepts one JPG/PNG up to 64 MiB; normalizes to 1920x1080 JPG. Saving does not publish or schedule. Uses the same deterministic conversion as the preview endpoint.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Optional for legacy clients; defaults to the normalized filename without extension.
                description:
                  type: string
                  maxLength: 2000
      responses:
        "400":
          description: Invalid image or metadata
        "403":
          description: Role cannot manage wallpapers
        "413":
          description: File exceeds the 64 MiB API limit or request exceeds the proxy limit
        "201":
          description: Wallpaper created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallpaper"
  /wallpapers/preview:
    post:
      tags: [Wallpapers]
      summary: Preview normalized wallpaper without persistence
      description: Operator or Administrator only. One non-empty JPG/PNG, maximum 64 MiB. Returns the same normalized image that a subsequent upload of that file produces. No database writes. Cache-Control is no-store.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Normalized 1920x1080 JPEG
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Missing, empty, unsupported or corrupt image
        "403":
          description: Role cannot manage wallpapers
        "413":
          description: Upload too large
  /wallpapers/{wallpaperId}:
    get:
      tags: [Wallpapers]
      summary: Get wallpaper detail including uploader and campaign references
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WallpaperId"
      responses:
        "200":
          description: Wallpaper detail (Operator or Administrator)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallpaper"
        "403":
          description: Role cannot access the Library
        "404":
          description: Wallpaper not found
    patch:
      tags: [Wallpapers]
      summary: Edit wallpaper title and description without changing image bytes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WallpaperId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                description:
                  type: string
                  nullable: true
                  maxLength: 2000
      responses:
        "200":
          description: Updated wallpaper; audit event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallpaper"
        "400":
          description: Invalid metadata (whitespace-only titles rejected)
        "403":
          description: Role cannot manage wallpapers
        "404":
          description: Wallpaper not found
    delete:
      tags: [Wallpapers]
      summary: Delete an eligible wallpaper
      description: Operator or Administrator only. Default wallpapers and DRAFT, SCHEDULED or ACTIVE campaign references block deletion. Soft-deletes and preserves historical references.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WallpaperId"
      responses:
        "204":
          description: Wallpaper deleted
        "400":
          description: Missing wallpaper, protected default, campaign reference, or concurrent mutation conflict
        "403":
          description: Role cannot manage wallpapers
  /wallpapers/{wallpaperId}/default:
    put:
      tags: [Wallpapers]
      summary: Choose default wallpaper (Administrator only)
      description: Changes only defaultWallpaperId and records wallpaper.default_changed. Does not interrupt an active campaign or publish immediately. Unknown/deleted targets are rejected.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WallpaperId"
      responses:
        "200":
          description: Default wallpaper selected
          content:
            application/json:
              schema:
                type: object
                required: [defaultWallpaperId]
                properties:
                  defaultWallpaperId:
                    type: string
                    format: uuid
        "403":
          description: Administrator role required
        "404":
          description: Wallpaper not found
  /wallpapers/{wallpaperId}/image:
    get:
      tags: [Wallpapers]
      summary: Get normalized wallpaper image bytes
      parameters:
        - $ref: "#/components/parameters/WallpaperId"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallpaper image stream
          headers:
            ETag:
              schema:
                type: string
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: >
        Protected endpoints return the standard Unauthorized response when the
        token is missing, expired, revoked, unknown, or belongs to an inactive user.
...
  parameters:
    WallpaperId:
      name: wallpaperId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Wallpaper:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        filename:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        resolution:
          type: string
        width:
          type: integer
        height:
          type: integer
        sizeBytes:
          type: integer
        checksumSha256:
          type: string
        mimeType:
          type: string
        imageUrl:
          type: string
        uploadedAt:
          type: string
          format: date-time
        isDefault:
          type: boolean
        uploadedBy:
          type: string
          description: Uploader username, included in list/detail responses
        usageStatus:
          type: string
          enum: [IN_USE, SCHEDULED, DRAFT]
          description: Legacy compatibility field; use campaigns for independent usage badges.
        campaigns:
          type: array
          description: All associated campaigns, including past use; included in list/detail responses
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              status:
                type: string
                enum: [DRAFT, SCHEDULED, ACTIVE, COMPLETED, CANCELLED]
              startDate:
                type: string
                format: date-time
                nullable: true
              endDate:
                type: string
                format: date-time
                nullable: true
...